app/requirements.txt
Django==4.2.5
psycopg2==2.9.7
gunicorn==21.2.0
django-phonenumber-field==7.1.0
phonenumberslite==8.13.19
